feat(directory-service): support embed of represented resource

+++ apps/directory-service/src/directory/model/resource.ts
@@ -1,10 +1,10 @@
-import { AdspId, assertAdspId, ServiceDirectory, TokenProvider } from '@abgov/adsp-service-sdk';
+import { AdspId, assertAdspId, ServiceDirectory } from '@abgov/adsp-service-sdk';
 import { DomainEvent, InvalidOperationError, NotFoundError } from '@core-services/core-common';
 import { property } from 'lodash';
-import { Resource, ResourceTypeConfiguration } from '../types';
 import axios from 'axios';
 import { Logger } from 'winston';
 import { DirectoryRepository } from '../repository';
+import { Resource, ResourceTypeConfiguration } from '../types';
 
 export class ResourceType {
   public type: string;
@@ -18,7 +18,6 @@ export class ResourceType {
   constructor(
     private logger: Logger,
     private directory: ServiceDirectory,
-    private tokenProvider: TokenProvider,
     private repository: DirectoryRepository,
     { type, matcher, namePath, descriptionPath, deleteEvent }: ResourceTypeConfiguration
   ) {
@@ -36,7 +35,7 @@ export class ResourceType {
     return this.matcher.test(urn.resource);
   }
 
-  public async resolve(resource: Resource): Promise<Resource> {
+  public async resolve(token: string, resource: Resource, deleteNotFound = false): Promise<Resource> {
     if (!this.matches(resource?.urn)) {
       throw new InvalidOperationError(`Resource type '${this.type}' not matched to resource: ${resource.urn}`);
     }
@@ -46,8 +45,6 @@ export class ResourceType {
       if (!resourceUrl) {
         throw new NotFoundError(`Failed to lookup URL for resource: ${resource.urn}.`);
       }
-
-      const token = await this.tokenProvider.getAccessToken();
 
       this.logger.debug(`Retrieving resource ${resource.urn} from: ${resourceUrl}...`, {
         context: 'ResourceType',
@@ -66,22 +63,25 @@ export class ResourceType {
 
       const name = this.nameGetter(data) || resource.name;
       const description = this.descriptionGetter(data) || resource.description;
+      if (name !== resource.name || description !== resource.description) {
+        resource = await this.repository.saveResource({ ...resource, name, description, type: this.type });
+      }
 
-      return await this.repository.saveResource({ ...resource, name, description, type: this.type });
+      return { ...resource, data };
     } catch (err) {
-      if (axios.isAxiosError(err) && err.response?.status === 404) {
+      if (axios.isAxiosError(err) && err.response?.status === 404 && deleteNotFound) {
         // If there's a 404, then the resource doesn't exist. This can happen if the resource was deleted before being resolved.
         // Delete the missing resource for consistency.
         await this.repository.deleteResource(resource);
-        // undefined is returned.
+        // Undefined is returned after delete;
         return;
       } else {
         this.logger.warn(`Error encountered resolving resource ${resource.urn}: ${err}`, {
           context: 'ResourceType',
           tenant: resource.tenantId?.toString(),
         });
-
-        throw err;
+        // Original resource is returned.
+        return resource;
       }
     }
   }
